{{cookiecutter.__project_kebab}}/pyproject.toml
[build-system]
requires = ["poetry-core>=1.7.1"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "{{cookiecutter.__project_kebab}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.author_full_name}} <{{cookiecutter.author_email}}>"]


[tool.poetry.dependencies]
# Server deps
python = "^3.12"
uvloop = "^0.19.0"
uvicorn = "^0.24.0"
fastapi = "^0.105.0"
pydantic = "^2.5.2"
pyyaml = "^6.0.1"
sentry-sdk = "^1.39.1"
starlette-exporter = "^0.17.1"



[tool.poetry.group.test]
optional = true


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"


[tool.poetry.group.lint]
optional = true


[tool.poetry.group.lint.dependencies]
pyright = "^1.1.341"
pylint = "^3.0.3"


[[tool.poetry.source]]
name = "distribusion"
url = "https://gitlab.com/api/v4/groups/15793688/-/packages/pypi/simple"
priority = "supplemental"


[tool.isort]
profile = "black"


[tool.pylint.messages_control]
disable = "C0115, R0903, C0116, R1734, R1735, E0401, E0611, R0913, R0902"


[tool.pylint.format]
max-line-length = "88"


[tool.black]
include = "{{cookiecutter.__project_slug}}"


# At the moment, this might break the CI
# + In general, pyright and env are to be configured by user
# [tool.pyright]
# include = ["{{cookiecutter.__project_slug}}"]
# venvPath = "."
# venv = ".venv"


[tool.coverage.run]
source = ["{{cookiecutter.__project_slug}}"]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "strict"
